Extract out gap prop
